#program-added

+++ perimeter-calculator.py
@@ -1,6 +1,6 @@
 print("this program will help you to find the perimeter of a poligon")
 
-l=int(input("Enter the length of a poligon "))
-b=int(input("Enter the bredth of a poligon "))
+l=float(input("Enter the length of a poligon "))
+b=float(input("Enter the bredth of a poligon "))
 
 print('The perimeter of the poligon is ',2*l+2*b)
